apps/cupid/backend/pyproject.toml
[project]
name = "cupid-backend"
version = "0.1.0"
description = "Cupid Deluxe matchmaking game backend with ChatKit"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.114.1,<0.116",
    "uvicorn[standard]>=0.36,<0.37",
    "openai>=1.40",
    "openai-chatkit>=1.1.2,<2",
    "openai-agents>=0.0.16",
    "pyyaml>=6.0",
    "jinja2>=3.1",
    "python-dotenv>=1.0",
]

[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"
